% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_sample_id.R
\name{assign_sample_id}
\alias{assign_sample_id}
\title{Assign Sample IDs and Annotate Tumor Information}
\usage{
assign_sample_id(
  this_data = NULL,
  start_id = NULL,
  lab_id_col = "lab_id",
  personal_id_col = "personal_id",
  date_col = "date_of_sample",
  verbose = TRUE
)
}
\arguments{
\item{this_data}{A data frame containing the input data.}

\item{start_id}{An integer specifying the starting number for the sample IDs. This is required.}

\item{lab_id_col}{A string specifying the column name for lab IDs. Default is `"lab_id"`.}

\item{personal_id_col}{A string specifying the column name for personal IDs. Default is `"personal_id"`.}

\item{date_col}{A string specifying the column name for the date of sample collection. Default is `"date_of_sample"`.}

\item{verbose}{A boolean indicating whether to print summary statistics about the data. Default is `TRUE`.}
}
\value{
A data frame with the following columns:
\describe{
  \item{lab_id}{The original lab ID column.}
  \item{personal_id}{The original personal ID column.}
  \item{date_of_sample}{The original date of sample collection column.}
  \item{sample_id}{The formatted sample ID column, which includes unique IDs, replicate numbers, and tumor annotations.}
}
}
\description{
This function assigns unique sample IDs, annotates duplicates, and tracks tumor occurrences for each patient.
It also provides an option to print summary statistics about the data.
}
\details{
The function performs the following steps:
\enumerate{
  \item Validates that the specified column names exist in the input data.
  \item Assigns unique sample IDs starting from the specified `start_id`.
  \item Annotates duplicate samples with a `sample_rep` column.
  \item Tracks tumor occurrences with a `tumor_n` column, assigning unique letters for distinct tumors.
  \item Replaces "A" in the `tumor_n` column with `NA` and adjusts the `sample_rep` column to start from 0.
  \item Creates a formatted sample ID by concatenating `sample_id`, `sample_rep`, and `tumor_n`.
  \item Optionally prints summary statistics, including the number of unique samples, replicates, and tumors.
}
}
\examples{
# Example data
test_data <- tibble(
  lab_id = c("45KF98987", "13KF85140", "13KF85140", "82KF36719"),
  personal_id = c("19037644-5806", "19053341-4128", "19053341-4128", "19053341-4128"),
  date_of_sample = c("2021-08-20", "2018-09-15", "2018-09-15", "2024-03-12")
)

# Run the function
result <- assign_sample_id(
  this_data = test_data,
  start_id = 100,
  lab_id_col = "lab_id",
  personal_id_col = "personal_id",
  date_col = "date_of_sample",
  verbose = TRUE
)

}
