% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate a Test Data Set with Controlled Sample Distribution}
\usage{
generate_test_data(num_samples, num_unique_personal_ids, num_unique_dates)
}
\arguments{
\item{num_samples}{Integer. The total number of samples to generate.}

\item{num_unique_personal_ids}{Integer. The number of unique personal IDs to generate.}

\item{num_unique_dates}{Integer. The number of unique dates to generate.}
}
\value{
A data frame with the following columns:
\describe{
  \item{lab_id}{Character. A unique lab ID for each sample.}
  \item{personal_id}{Character. A unique personal ID for each sample.}
  \item{date_of_sample}{Date. The date associated with each sample.}
}
}
\description{
This function generates a test data set with a specified number of samples, unique personal IDs,
unique dates, and samples per unique combination of personal ID and date. It ensures that the
total number of samples matches the specified `num_samples`.
}
\examples{
#' #load pacakges
library(tibble, dplyr)

# Generate a test data set with 100 samples, 10 unique personal IDs, and 5 unique dates
test_data <- generate_test_data(
  num_samples = 100,
  num_unique_personal_ids = 10,
  num_unique_dates = 5
)
head(test_data)

}
